pyproject.toml
[project]
name = "cardboardcompanion"
version = "0.1.0"
description = ""
authors = [
    {name = "Tres Sims",email = "tres.sims95@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13.0"
dependencies = [
    "pydantic (>=2.11.3,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "click (>=8.1.8,<9.0.0)",
    "logging (>=0.4.9.6,<0.5.0.0)",
    "py-cord (>=2.6.1,<3.0.0)",
]

[tool.poetry]
packages = [{include = "cardboardcompanion", from = "src"}]

[tool.poetry.scripts]
ccompanion = "cardboardcompanion.cmd:ccompanion"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
